+++ src/tab/tab-model.d.ts
@@ -177,7 +177,7 @@ export interface TabModel extends ComponentModel{
      * - None - Based on height property given, content panel height is set.
      * - Auto - Tallest panel height of given Tab contents is set to all other panels.
      * - Content - Based on the corresponding content height, content panel height is set.
-     * - Fill - Based on the parent height, content panel hight is set.
+     * - Fill - Based on the parent height, content panel height is set.
      * @default 'Content'
      */
     heightAdjustMode?: heightStyles;